cosmetic: compile warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@28216 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDCodecs/Video/DVDVideoCodecCrystalHD.cpp
@@ -162,7 +162,7 @@ int CDVDVideoCodecCrystalHD::Decode(BYTE *pData, int iSize, double dts, double p
   if (!m_DropPictures)
   {
     if (m_Duration > 0.0)
-      Sleep(m_Duration/2000.0);
+      Sleep((DWORD)(m_Duration/2000.0));
     else
       Sleep(20);
   }
